[WEB-2413] chore: admin application restructuring.
