+++ admin/ee/components/common/index.ts
@@ -1,0 +1,1 @@
+export * from "ce/components/common";
