+++ admin/ee/store/root.store.ts
@@ -1,0 +1,1 @@
+export * from "ce/store/root.store";
